backend/pyproject.toml
[project]
name = "bookcompanion"
version = "0.1.0"
description = "Personal book summarization and knowledge extraction tool"
requires-python = ">=3.12"
dependencies = [
    # CLI
    "typer>=0.15.0",
    "rich>=13.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.4.0",
    # Config
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    # Parsing
    "ebooklib>=0.18",
    "markdownify>=0.14.0",
    "pymupdf4llm>=0.0.17",
    # Embeddings & HTTP
    "httpx>=0.28.0",
    # Clipboard (for --copy flag)
    "pyperclip>=1.9.0",
    # Prompts
    "jinja2>=3.1.0",
    # Logging
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
bookcompanion = "app.cli.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
markers = [
    "integration_llm: tests that make real LLM API calls (slow, require Claude CLI)",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
